\name{print.vbglmss}
\alias{print.vbglmss}
\title{Variational approximation to Spike and Slab GLM.}
\usage{
  print.vbglmss(x, ...)
}
\arguments{
  \item{y}{Test param y for w}

  \item{X}{X param}

  \item{Z}{param Z}
}
\value{
  A list with cool stuff.
}
\description{
  Variational approximation to Spike and Slab GLM.
}

