\name{vbglmss.fit.logistic.X}
\alias{vbglmss.fit.logistic.X}
\title{Fit logistic regression with only SS coefficients.}
\usage{
  vbglmss.fit.logistic.X(y, X, prior, cat2, eps = 0.01,
    ...)
}
\arguments{
  \item{y}{binary response.}

  \item{X}{the SS covariates.}

  \item{prior}{a list of priors}

  \item{cat2}{printer}

  \item{eps}{the approximation loop convergence threshold}
}
\description{
  Fit logistic regression with only SS coefficients.
}

