\name{vbglmss.fit.logistic.Z}
\alias{vbglmss.fit.logistic.Z}
\title{Fit logistic with only gaussian coef covariates}
\usage{
  vbglmss.fit.logistic.Z(y, Z, prior, cat2, eps, ...)
}
\arguments{
  \item{y}{binary response.}

  \item{Z}{the covariates.}

  \item{prior}{a list of priors}

  \item{cat2}{printer}

  \item{eps}{the approximation loop convergence threshold}
}
\description{
  Fit logistic with only gaussian coef covariates
}

